+++ src/api/c/cvc5.cpp
@@ -3745,7 +3745,7 @@ Cvc5Result cvc5_result_copy(Cvc5Result result)
   Cvc5Result res = nullptr;
   CVC5_CAPI_TRY_CATCH_BEGIN;
   CVC5_CAPI_CHECK_RESULT(result);
-  res = result->d_tm->copy(result);
+  res = result->copy();
   CVC5_CAPI_TRY_CATCH_END;
   return res;
 }
@@ -3754,7 +3754,7 @@ void cvc5_result_release(Cvc5Result result)
 {
   CVC5_CAPI_TRY_CATCH_BEGIN;
   CVC5_CAPI_CHECK_RESULT(result);
-  result->d_tm->release(result);
+  result->release();
   CVC5_CAPI_TRY_CATCH_END;
 }
 
@@ -3871,7 +3871,7 @@ Cvc5SynthResult cvc5_synth_result_copy(Cvc5SynthResult result)
   Cvc5SynthResult res = nullptr;
   CVC5_CAPI_TRY_CATCH_BEGIN;
   CVC5_CAPI_CHECK_SYNTH_RESULT(result);
-  res = result->d_tm->copy(result);
+  res = result->copy();
   CVC5_CAPI_TRY_CATCH_END;
   return res;
 }
@@ -3880,7 +3880,7 @@ void cvc5_synth_result_release(Cvc5SynthResult result)
 {
   CVC5_CAPI_TRY_CATCH_BEGIN;
   CVC5_CAPI_CHECK_SYNTH_RESULT(result);
-  result->d_tm->release(result);
+  result->release();
   CVC5_CAPI_TRY_CATCH_END;
 }
 
@@ -4022,7 +4022,7 @@ const Cvc5Proof* cvc5_proof_get_children(Cvc5Proof proof, size_t* size)
   auto children = proof->d_proof.getChildren();
   for (auto& p : children)
   {
-    res.push_back(proof->d_tm->export_proof(p));
+    res.push_back(proof->export_proof(p));
   }
   *size = res.size();
   CVC5_CAPI_TRY_CATCH_END;
@@ -4093,7 +4093,7 @@ Cvc5Proof cvc5_proof_copy(Cvc5Proof proof)
   Cvc5Proof res = nullptr;
   CVC5_CAPI_TRY_CATCH_BEGIN;
   CVC5_CAPI_CHECK_PROOF(proof);
-  res = proof->d_tm->copy(proof);
+  res = proof->copy();
   CVC5_CAPI_TRY_CATCH_END;
   return res;
 }
@@ -4102,7 +4102,7 @@ void cvc5_proof_release(Cvc5Proof proof)
 {
   CVC5_CAPI_TRY_CATCH_BEGIN;
   CVC5_CAPI_CHECK_PROOF(proof);
-  proof->d_tm->release(proof);
+  proof->release();
   CVC5_CAPI_TRY_CATCH_END;
 }
 
@@ -4214,7 +4214,7 @@ Cvc5Grammar cvc5_grammar_copy(Cvc5Grammar grammar)
   Cvc5Grammar res = nullptr;
   CVC5_CAPI_TRY_CATCH_BEGIN;
   CVC5_CAPI_CHECK_GRAMMAR(grammar);
-  res = grammar->d_tm->copy(grammar);
+  res = grammar->copy();
   CVC5_CAPI_TRY_CATCH_END;
   return res;
 }
@@ -4223,7 +4223,7 @@ void cvc5_grammar_release(Cvc5Grammar grammar)
 {
   CVC5_CAPI_TRY_CATCH_BEGIN;
   CVC5_CAPI_CHECK_GRAMMAR(grammar);
-  grammar->d_tm->release(grammar);
+  grammar->release();
   CVC5_CAPI_TRY_CATCH_END;
 }
 
@@ -4846,7 +4846,7 @@ Cvc5Result cvc5_check_sat(Cvc5* cvc5)
   Cvc5Result res = nullptr;
   CVC5_CAPI_TRY_CATCH_BEGIN;
   CVC5_CAPI_CHECK_NOT_NULL(cvc5);
-  res = cvc5->d_tm->export_result(cvc5->d_solver.checkSat());
+  res = cvc5->export_result(cvc5->d_solver.checkSat());
   CVC5_CAPI_TRY_CATCH_END;
   return res;
 }
@@ -4865,8 +4865,7 @@ Cvc5Result cvc5_check_sat_assuming(Cvc5* cvc5,
     CVC5_CAPI_CHECK_TERM_AT_IDX(assumptions, i);
     cassumptions.push_back(assumptions[i]->d_term);
   }
-  res =
-      cvc5->d_tm->export_result(cvc5->d_solver.checkSatAssuming(cassumptions));
+  res = cvc5->export_result(cvc5->d_solver.checkSatAssuming(cassumptions));
   CVC5_CAPI_TRY_CATCH_END;
   return res;
 }
@@ -5173,7 +5172,7 @@ const Cvc5Term* cvc5_get_timeout_core(Cvc5* cvc5,
   CVC5_CAPI_CHECK_NOT_NULL(size);
   res.clear();
   auto ccore = cvc5->d_solver.getTimeoutCore();
-  *result = cvc5->d_tm->export_result(ccore.first);
+  *result = cvc5->export_result(ccore.first);
   for (const auto& t : ccore.second)
   {
     res.push_back(cvc5->d_tm->export_term(t));
@@ -5203,7 +5202,7 @@ const Cvc5Term* cvc5_get_timeout_core_assuming(Cvc5* cvc5,
   CVC5_CAPI_CHECK_NOT_NULL(rsize);
   res.clear();
   auto ccore = cvc5->d_solver.getTimeoutCoreAssuming(cassumptions);
-  *result = cvc5->d_tm->export_result(ccore.first);
+  *result = cvc5->export_result(ccore.first);
   for (const auto& t : ccore.second)
   {
     res.push_back(cvc5->d_tm->export_term(t));
@@ -5224,7 +5223,7 @@ const Cvc5Proof* cvc5_get_proof(Cvc5* cvc5, Cvc5ProofComponent c, size_t* size)
       cvc5->d_solver.getProof(static_cast<cvc5::modes::ProofComponent>(c));
   for (const auto& p : proofs)
   {
-    res.push_back(cvc5->d_tm->export_proof(p));
+    res.push_back(cvc5->export_proof(p));
   }
   *size = proofs.size();
   CVC5_CAPI_TRY_CATCH_END;
@@ -5688,8 +5687,7 @@ Cvc5Grammar cvc5_mk_grammar(Cvc5* cvc5,
   {
     csymbols.push_back(symbols[i]->d_term);
   }
-  res = cvc5->d_tm->export_grammar(
-      cvc5->d_solver.mkGrammar(cbound_vars, csymbols));
+  res = cvc5->export_grammar(cvc5->d_solver.mkGrammar(cbound_vars, csymbols));
   CVC5_CAPI_TRY_CATCH_END;
   return res;
 }
@@ -5821,7 +5819,7 @@ Cvc5SynthResult cvc5_check_synth(Cvc5* cvc5)
   Cvc5SynthResult res = nullptr;
   CVC5_CAPI_TRY_CATCH_BEGIN;
   CVC5_CAPI_CHECK_NOT_NULL(cvc5);
-  res = cvc5->d_tm->export_synth_result(cvc5->d_solver.checkSynth());
+  res = cvc5->export_synth_result(cvc5->d_solver.checkSynth());
   CVC5_CAPI_TRY_CATCH_END;
   return res;
 }
@@ -5831,7 +5829,7 @@ Cvc5SynthResult cvc5_check_synth_next(Cvc5* cvc5)
   Cvc5SynthResult res = nullptr;
   CVC5_CAPI_TRY_CATCH_BEGIN;
   CVC5_CAPI_CHECK_NOT_NULL(cvc5);
-  res = cvc5->d_tm->export_synth_result(cvc5->d_solver.checkSynthNext());
+  res = cvc5->export_synth_result(cvc5->d_solver.checkSynthNext());
   CVC5_CAPI_TRY_CATCH_END;
   return res;
 }